Add files via upload
This is the First attempt at making a game just text based so you have to use Visual studios (Preferably 2017) to run it. I have a lot of work still to do on it and I have already made a new version of it that I will be posting hopefully today.